add some doc generation

+++ support/gen_docs.cr
@@ -1,6 +1,6 @@
 #!/usr/bin/env crystal
 
-require "./helper"
+require "file_utils"
 require "../src/kube-client/version"
 
 DOCS_DIR     = "docs"
@@ -9,8 +9,10 @@ LASTEST      = Kube::Client::VERSION
 PROJ_NAME    = "kube-client.cr"
 
 def for_each_version
+  return unless Dir.exists?(VERSIONS_DIR)
   Dir.open(VERSIONS_DIR).each_child do |version|
-    yield version.rchomp(".cr"), nil, "#{version.rchomp(".cr")}.0"
+    next unless version =~ /^v\d/
+    yield version.chomp(".cr"), nil, "#{version.chomp(".cr")}.0"
   end
 end
 
@@ -41,7 +43,7 @@ def generate_docs_for(prefix, version)
   rel_ver = get_release_version
   docs_dir = File.join(".", DOCS_DIR, rel_ver, prefix)
   version_file = File.join(".", VERSIONS_DIR, "#{prefix}.cr")
-
+  return unless File.exists?(version_file)
   FileUtils.rm_rf File.join(docs_dir) if Dir.exists?(docs_dir)
   FileUtils.mkdir_p(docs_dir) unless Dir.exists?(docs_dir)
 
@@ -49,14 +51,21 @@ def generate_docs_for(prefix, version)
 end
 
 def _generate_docs(version_file, docs_dir, rel_ver, git_commit, api_ver)
+  # lib/k8s/src/versions
+
   args = [
     "doc",
-    version_file,
     "--project-name", PROJ_NAME,
     "--output", docs_dir,
     "--project-version", rel_ver,
     "--source-refname", git_commit,
   ]
+
+  Dir.glob("./lib/k8s/src/k8s/*.cr").each do |file|
+    args << file
+  end
+  args << File.join(".", "lib/k8s/src/versions", File.basename(version_file))
+  args << version_file
 
   system "crystal", args
 
@@ -87,13 +96,13 @@ def generate_release_docs
     generate_release_docs_for(tag[0].lchop('v'), tag[1])
   end
 
-  `git checkout #{current_ref}`
+  `git checkout master`
   generate_version_list(File.join(".", DOCS_DIR), docs, "K8S Releases")
 end
 
 def generate_release_docs_for(tag, commit)
   `git checkout #{commit}`
-
+  `shards install`
   versions = [] of String
   for_each_version do |prefix, _, version|
     generate_docs_for(prefix, version)
@@ -104,9 +113,10 @@ def generate_release_docs_for(tag, commit)
 end
 
 def generate_version_list(docs_dir, docs, title = nil)
+  return unless Dir.exists?(docs_dir)
   index = File.join(docs_dir, "index.html")
   File.open(index, "w") do |f|
-    f.puts gen_index(title, docs)
+    f.puts gen_index(title, docs.sort.reverse!)
   end
 end
